+++ src/Symfony/Component/Messenger/Transport/Sender/Locator/SenderLocator.php
@@ -9,11 +9,11 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\Messenger\Asynchronous\Routing;
+namespace Symfony\Component\Messenger\Transport\Sender\Locator;
 
 use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\Exception\RuntimeException;
-use Symfony\Component\Messenger\Transport\SenderInterface;
+use Symfony\Component\Messenger\Transport\Sender\SenderInterface;
 
 /**
  * @author Fabien Potencier <fabien@symfony.com>